base仓告警清理

+++ services/netmanagernative/src/manager/interface_manager.cpp
@@ -505,7 +505,7 @@ int32_t InterfaceManager::AssembleArp(const std::string &ipAddr, const std::stri
         return NETMANAGER_ERR_OPERATION_FAILED;
     }
 
-    req.arp_flags |= ATF_COM;
+    static_cast<unsigned int>(req.arp_flags)|= ATF_COM;
 
     return NETMANAGER_SUCCESS;
 }
